Updated 204 error
Updated 204 error description per request from Charles and Kara:
As we discussed yesterday, this appears to be an API issue in that distances of 500+ meters are returned as an error message: "Input parameters are invalid or nearest roadway is >500 meters from specified latitude/longitude". +lisa@renci.org: I think we need to fix this and return distances of >=500 meters as such, with invalid input parameters differentiated as a true error, if that makes sense.